+++ core/src/main/java/org/infinispan/loaders/AbstractCacheLoaderConfig.java
@@ -1,13 +1,20 @@
 package org.infinispan.loaders;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.util.Properties;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlTransient;
 
 import org.infinispan.CacheException;
+import org.infinispan.config.AbstractNamedCacheConfigurationBean;
 import org.infinispan.config.ConfigurationBeanVisitor;
 import org.infinispan.config.ConfigurationDocRef;
-import org.infinispan.config.PluggableConfigurationComponent;
+import org.infinispan.config.parsing.XmlConfigHelper;
+import org.infinispan.util.TypedProperties;
 
 /**
  * Abstract base class for CacheLoaderConfigs.
@@ -19,7 +26,7 @@ import org.infinispan.config.PluggableConfigurationComponent;
  * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-public class AbstractCacheLoaderConfig extends PluggableConfigurationComponent implements CacheLoaderConfig {
+public class AbstractCacheLoaderConfig extends AbstractNamedCacheConfigurationBean implements CacheLoaderConfig {
 
    /** The serialVersionUID */
    private static final long serialVersionUID = -4303705423800914433L;
@@ -27,6 +34,31 @@ public class AbstractCacheLoaderConfig extends PluggableConfigurationComponent i
    @XmlAttribute(name="class")
    @ConfigurationDocRef(name="class", bean=AbstractCacheLoaderConfig.class,targetElement="setCacheLoaderClassName")
    protected String cacheLoaderClassName;
+     
+   @XmlTransient
+   protected TypedProperties properties = EMPTY_PROPERTIES;
+
+   public Properties getProperties() {
+      return properties;
+   }
+   
+   public void setProperties(Properties properties) {
+      testImmutability("properties");
+      this.properties = toTypedProperties(properties);
+   }
+
+   public void setProperties(String properties) throws IOException {
+      if (properties == null) return;
+
+      testImmutability("properties");
+      // JBCACHE-531: escape all backslash characters
+      // replace any "\" that is not preceded by a backslash with "\\"
+      properties = XmlConfigHelper.escapeBackslashes(properties);
+      ByteArrayInputStream is = new ByteArrayInputStream(properties.trim().getBytes("ISO8859_1"));
+      this.properties = new TypedProperties();
+      this.properties.load(is);
+      is.close();
+   }
 
    public String getCacheLoaderClassName() {
       return cacheLoaderClassName;
